% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosave_module.R
\name{autosave_module_ui}
\alias{autosave_module_ui}
\title{Autosave module UI}
\usage{
autosave_module_ui(id)
}
\arguments{
\item{id}{Module identifier.}
}
\value{
A Shiny UI fragment.
}
\description{
Autosave module UI
}
\keyword{internal}
