% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_main_server.R
\name{build_group_defs}
\alias{build_group_defs}
\title{Build group definitions from a list of group names}
\usage{
build_group_defs(group_list)
}
\description{
Build group definitions from a list of group names
}
\keyword{internal}
