% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_main_server.R
\name{build_taxon_choices}
\alias{build_taxon_choices}
\title{Build choices for the main taxon selectize input}
\usage{
build_taxon_choices(taxonomy, show_out_of_region = FALSE)
}
\description{
Build choices for the main taxon selectize input
}
\keyword{internal}
