% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_export.R
\name{calculate_corixid_ratio_static}
\alias{calculate_corixid_ratio_static}
\title{Calculate corixid ratio without reactives}
\usage{
calculate_corixid_ratio_static(
  section_data,
  group_totals,
  taxonomy,
  beetles_section,
  bugs_section
)
}
\arguments{
\item{section_data}{Section data list.}

\item{group_totals}{Named vector of group totals.}

\item{taxonomy}{Taxonomy data frame.}

\item{beetles_section}{Section ID for beetles.}

\item{bugs_section}{Section ID for bugs.}
}
\value{
Numeric corixid ratio.
}
\description{
Calculate corixid ratio without reactives
}
\keyword{internal}
