% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_metric_calculation.R
\name{calculate_metric_score}
\alias{calculate_metric_score}
\title{Calculate a metric score}
\usage{
calculate_metric_score(
  values,
  min_values,
  max_values,
  scale_factor = 10,
  inverse = FALSE
)
}
\arguments{
\item{values}{Metric values.}

\item{min_values}{Minimum values for scaling.}

\item{max_values}{Maximum values for scaling.}

\item{scale_factor}{Scaling factor.}

\item{inverse}{Whether to invert the scale.}
}
\value{
Numeric scores.
}
\description{
Calculate a metric score
}
\keyword{internal}
