% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_export.R
\name{calculate_metrics_from_sections}
\alias{calculate_metrics_from_sections}
\title{Calculate metrics from section data without reactives}
\usage{
calculate_metrics_from_sections(section_data, taxonomy, group_defs)
}
\arguments{
\item{section_data}{Named list of section data (from build_sections_from_autosave).}

\item{taxonomy}{Taxonomy data frame.}

\item{group_defs}{Group definitions data frame.}
}
\value{
Data frame of metric scores.
}
\description{
Calculate metrics from section data without reactives
}
\keyword{internal}
