% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_module.R
\name{download_module_server}
\alias{download_module_server}
\title{Download module server}
\usage{
download_module_server(
  id,
  selected_genera,
  shared_reactives,
  auto_save_interval = 30,
  auto_save_path = get_app_path("autosave_dir"),
  expose_data_reactive = NULL
)
}
\arguments{
\item{id}{Module identifier.}

\item{selected_genera}{Reactive values of selected genera.}

\item{shared_reactives}{Shared reactive values.}

\item{auto_save_interval}{Interval used for autosave (reserved for future use).}

\item{auto_save_path}{Path to autosave directory (reserved for future use).}

\item{expose_data_reactive}{Optional reactiveVal used to expose assembled data.}
}
\description{
Download module server
}
\keyword{internal}
