% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_module.R
\name{download_module_ui}
\alias{download_module_ui}
\title{Download module UI}
\usage{
download_module_ui(id)
}
\arguments{
\item{id}{Module identifier.}
}
\value{
A Shiny UI fragment.
}
\description{
Download module UI
}
\keyword{internal}
