% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{escape_latex}
\alias{escape_latex}
\title{Escape LaTeX characters}
\usage{
escape_latex(x)
}
\arguments{
\item{x}{Character vector to escape.}
}
\value{
Escaped character vector.
}
\description{
Escape LaTeX characters
}
\keyword{internal}
