% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_export.R
\name{generate_pdf_report}
\alias{generate_pdf_report}
\title{Generate a PDF report from template}
\usage{
generate_pdf_report(
  filepath,
  user_title,
  user_date,
  metric_scores,
  template_name
)
}
\arguments{
\item{filepath}{Output file path.}

\item{user_title}{Report title.}

\item{user_date}{Sampling date.}

\item{metric_scores}{Metric scores data frame.}

\item{template_name}{Name of the Rmd template file.}
}
\description{
Generate a PDF report from template
}
\keyword{internal}
