% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R
\name{get_app_path}
\alias{get_app_path}
\title{Retrieve a stored path}
\usage{
get_app_path(key)
}
\arguments{
\item{key}{Name of the path to retrieve.}
}
\value{
A character path.
}
\description{
Retrieve a stored path
}
\keyword{internal}
