% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R
\name{init_app_state}
\alias{init_app_state}
\title{Initialize application paths}
\usage{
init_app_state()
}
\value{
A list of paths used by the app.
}
\description{
Initialize application paths
}
\keyword{internal}
