% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroibi-package.R
\docType{package}
\name{macroibi-package}
\alias{macroibi}
\alias{macroibi-package}
\title{macroibi: Wetland Macroinvertebrate IBI Dashboard}
\description{
Launches the Macroinvertebrate Index of Biotic Integrity Shiny dashboard and supporting utilities for wetlands.
}
\author{
\strong{Maintainer}: Sam Swanson \email{sam.swanson@shakopeedakota.org}

Authors:
\itemize{
  \item Sam Swanson \email{sam.swanson@shakopeedakota.org}
}

}
\keyword{internal}
