% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{macroibi_server}
\alias{macroibi_server}
\title{Internal server}
\usage{
macroibi_server(taxonomy, group_list, demo_mode = FALSE)
}
\arguments{
\item{taxonomy}{Taxonomy data frame.}

\item{group_list}{List of taxonomic groups.}
}
\description{
Internal server
}
\keyword{internal}
