% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_main_server.R
\name{normalize_in_region_flag}
\alias{normalize_in_region_flag}
\title{Normalize in_region to a logical flag}
\usage{
normalize_in_region_flag(taxonomy)
}
\description{
Normalize in_region to a logical flag
}
\keyword{internal}
