% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_export.R
\name{null-coalescing-operator}
\alias{null-coalescing-operator}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand side value.}

\item{y}{Right-hand side value (default).}
}
\value{
x if not NULL, otherwise y.
}
\description{
Returns the left-hand side if not NULL, otherwise the right-hand side.
}
\keyword{internal}
