% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_export.R
\name{process_autosave_file}
\alias{process_autosave_file}
\title{Process a single autosave file and generate outputs}
\usage{
process_autosave_file(
  file,
  autosave_path,
  output_path,
  types_to_generate,
  taxonomy,
  group_defs
)
}
\arguments{
\item{file}{Autosave filename.}

\item{autosave_path}{Path to autosave directory.}

\item{output_path}{Path to output directory.}

\item{types_to_generate}{Character vector of output types.}

\item{taxonomy}{Taxonomy data frame.}

\item{group_defs}{Group definitions data frame.}
}
\value{
Data frame of results.
}
\description{
Process a single autosave file and generate outputs
}
\keyword{internal}
