% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_generate_tree.R
\name{rank_levels}
\alias{rank_levels}
\title{Default taxonomic rank levels used for tree layout}
\usage{
rank_levels
}
\description{
These are the ITIS-like rank labels used to order nodes from
root to tips when drawing taxonomic trees.
}
\keyword{internal}
