% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_download.R
\name{results_download_ui}
\alias{results_download_ui}
\title{Results download UI}
\usage{
results_download_ui(id, demo_mode)
}
\arguments{
\item{id}{Module identifier.}
}
\value{
A Shiny UI fragment.
}
\description{
Results download UI
}
\keyword{internal}
