% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_utils.R
\name{sanitize_group_id}
\alias{sanitize_group_id}
\title{Sanitize group ID for stable use in file names/IDs}
\usage{
sanitize_group_id(x)
}
\description{
Sanitize group ID for stable use in file names/IDs
}
\keyword{internal}
