% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_metrics.R
\name{server_metrics}
\alias{server_metrics}
\title{Metric scores server}
\usage{
server_metrics(
  id,
  selected_genera,
  taxonomy,
  unique_taxa_counts,
  group_totals,
  grand_total_observations,
  group_defs
)
}
\arguments{
\item{id}{Module identifier.}

\item{selected_genera}{Reactive values of selected genera.}

\item{taxonomy}{Taxonomy data frame.}

\item{unique_taxa_counts}{Reactive values of unique taxa counts.}

\item{group_totals}{Reactive values of group totals.}

\item{grand_total_observations}{Reactive expression for grand total observations.}
}
\value{
Reactive values for metric scores.
}
\description{
Metric scores server
}
\keyword{internal}
