% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_taxon_section.R
\name{server_taxon_section}
\alias{server_taxon_section}
\title{Taxon section server}
\usage{
server_taxon_section(
  id,
  group_name,
  taxonomy_data,
  selected_taxon,
  total_unique_taxa
)
}
\arguments{
\item{id}{Module identifier.}

\item{group_name}{Display name for the group.}

\item{taxonomy_data}{Taxonomy data frame.}

\item{selected_taxon}{Reactive value representing the selected taxon.}

\item{total_unique_taxa}{Reactive expression with the total unique taxa count.}
}
\value{
A list of reactive values for the section.
}
\description{
Taxon section server
}
\keyword{internal}
