% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_main_server.R
\name{sum_group_totals}
\alias{sum_group_totals}
\title{Sum group totals with a minimum length requirement}
\usage{
sum_group_totals(counts, expected_groups)
}
\description{
Sum group totals with a minimum length requirement
}
\keyword{internal}
