% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{taxonomic_tree_server}
\alias{taxonomic_tree_server}
\title{Taxonomic tree server}
\usage{
taxonomic_tree_server(id, selected_genera, taxonomy_df)
}
\arguments{
\item{id}{Module identifier.}

\item{selected_genera}{Reactive for the current group's selected genera.
This should be a reactive that returns an object with a \verb{$data} field,
where \verb{$data} is a list of taxon rows (each row a list with \code{tsn}, etc.).}

\item{taxonomy_df}{Reactive taxonomy data frame.}
}
\description{
Taxonomic tree server
}
\keyword{internal}
