% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{taxonomic_tree_ui}
\alias{taxonomic_tree_ui}
\title{Taxonomic tree UI}
\usage{
taxonomic_tree_ui(id)
}
\arguments{
\item{id}{Module identifier.}
}
\value{
A Shiny UI fragment.
}
\description{
Taxonomic tree UI
}
\keyword{internal}
