% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_metrics.R
\name{ui_metric_scores}
\alias{ui_metric_scores}
\title{Metric scores UI}
\usage{
ui_metric_scores(id)
}
\arguments{
\item{id}{Module identifier.}
}
\value{
A Shiny UI fragment.
}
\description{
Metric scores UI
}
\keyword{internal}
