% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_taxon_section.R
\name{ui_taxon_section}
\alias{ui_taxon_section}
\title{Taxon section UI}
\usage{
ui_taxon_section(id, group_name, color)
}
\arguments{
\item{id}{Module identifier.}

\item{group_name}{Display name for the group.}

\item{color}{Color used for styling.}
}
\value{
A Shiny UI fragment.
}
\description{
Taxon section UI
}
\keyword{internal}
