% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_module.R
\name{upload_module_server}
\alias{upload_module_server}
\title{Upload module server}
\usage{
upload_module_server(
  id,
  taxonomy,
  selected_genera,
  shared_reactives,
  toggle_state,
  group_defs
)
}
\arguments{
\item{id}{Module identifier.}

\item{taxonomy}{Taxonomy data frame.}

\item{selected_genera}{Reactive values of selected genera.}

\item{shared_reactives}{Shared reactive values.}

\item{toggle_state}{Reactive value controlling editing state.}

\item{group_defs}{Data frame with columns section_id, group_id, group_name
mapping groups to section modules.}
}
\description{
Upload module server
}
\keyword{internal}
