% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_module.R
\name{upload_module_ui}
\alias{upload_module_ui}
\title{Upload module UI}
\usage{
upload_module_ui(id)
}
\arguments{
\item{id}{Module identifier.}
}
\value{
A Shiny UI fragment.
}
\description{
Upload module UI
}
\keyword{internal}
