% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_metric_calculation.R
\name{calculate_corixids_ratio}
\alias{calculate_corixids_ratio}
\title{Calculate the Corixids ratio metric}
\usage{
calculate_corixids_ratio(selected_genera, group_totals, taxonomy)
}
\arguments{
\item{selected_genera}{Reactive values of selected taxa.}

\item{group_totals}{Reactive values of group totals.}

\item{taxonomy}{Taxonomy data frame.}
}
\value{
Numeric ratio.
}
\description{
Calculate the Corixids ratio metric
}
\keyword{internal}
