% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R
\docType{package}
\name{macroibi-package}
\alias{macroibi}
\alias{macroibi-package}
\title{Internal application state}
\description{
Utilities to manage package paths and user data directories.
}
\author{
\strong{Maintainer}: Sam Swanson \email{sam.swanson@example.com}

}
\keyword{internal}
