% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R, R/macroibi-package.R
\docType{package}
\name{macroibi-package}
\alias{macroibi-package}
\title{Internal application state}
\description{
Utilities to manage package paths and user data directories.

Launches the Macroinvertebrate Index of Biotic Integrity Shiny dashboard and supporting utilities for wetlands.
}
\author{
\strong{Maintainer}: Sam Swanson \email{sam.swanson@example.com}

}
\keyword{internal}
