% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_download.R
\name{results_download_server}
\alias{results_download_server}
\title{Results download server}
\usage{
results_download_server(
  id,
  metric_scores,
  shared_reactives,
  selected_genera,
  taxonomy,
  quality_class,
  metric_save_path = get_app_path("metric_autosave_dir"),
  group_defs
)
}
\arguments{
\item{id}{Module identifier.}

\item{metric_scores}{Reactive values for metric scores.}

\item{shared_reactives}{Shared reactive values.}

\item{selected_genera}{Reactive values of selected genera.}

\item{taxonomy}{Taxonomy data frame.}

\item{metric_save_path}{Path where metric autosave files are stored.}

\item{group_defs}{Data frame with columns section_id, group_id, group_name
mapping section modules to stable groups.}
}
\description{
Results download server
}
\keyword{internal}
