% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{taxonomic_tree_server}
\alias{taxonomic_tree_server}
\title{Taxonomic tree server}
\usage{
taxonomic_tree_server(id, selected_genera, taxonomy_df)
}
\arguments{
\item{id}{Module identifier.}

\item{selected_genera}{Reactive values of selected genera.}

\item{taxonomy_df}{Reactive taxonomy data frame.}
}
\description{
Taxonomic tree server
}
\keyword{internal}
